ai-tagger/pyproject.toml
[project]
name = "ai-tagger"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.41.0",
]
